Add entity.hpp and ECS infrastructure

+++ Hazel/Src/Hazel/Core/LayerStack.h
@@ -14,7 +14,7 @@ namespace Hazel
 		void PopOverlay(std::shared_ptr<Layer> overlay);
 		std::shared_ptr<Layer> PopLayer();
 		std::shared_ptr<Layer> GetLayer(uint32_t id) { return m_Stack[id]; }
-		uint32_t GetLayerCnt() { return m_Stack.size(); }
+		uint32_t GetLayerCnt() { return (uint32_t)m_Stack.size(); }
 
 		std::vector<std::shared_ptr<Layer>>::iterator begin() {	return m_Stack.begin(); }
 		std::vector<std::shared_ptr<Layer>>::iterator end() {	return m_Stack.end(); }
